Shrink `Modifier` and `Inventory`.
By using `Box<[Slot]>` instead of `Vec<Slot>` in `Inventory`, we can
reduce the size of `Inventory` to 2 pointers and therefore `Modifier`
to 3 pointers. The largest variant of `Modifier` is now `Composite`, but
the `Composite` struct has niche-bearing fields, so `Modifier` is the
same size as `Composite` now.

I'm doing this not because of any known performance issue, but because
it seems to me to be a good idea to keep the size of `Modifier` under
control so that blocks' `Vec<Modifier>` is compact relative to how much
information it's actually storing.