content: Reduce size of `lighting_bench_space()`'s future.
I’m attempting to make this code able to run with a very small stack,
which means making the future small so that allocating it isn't a
problem. It’d be more practical to make a non-`async` version, but at
least I’m learning things about async block code generation.