Split out new crate `all-is-cubes-render`.
It contains `HeadlessRenderer` and all the other code from `camera` and
`raytracer` that I could manage to move. The specific rationale for this
is that I want to add more esoteric raytracer features, and they
shouldn't add to the fundamental crate's build costs.

More broadly, it continues to be the case that the project as a whole
takes an awful lot of time to build, and any split that produces extra
build parallelism is nice. Unfortunately, `all-is-cubes-ui` needs
`StandardCameras`, so the critical path isn't much shorter.

Future work:

* If we get rid of the raytracing algorithms' dependency on
  `GraphicsOptions`, then we can move `GraphicsOptions` and friends
  actually out of `all-is-cubes`. Or maybe we should give up on
  raytracing in unit tests entirely and just have the tests dump
  serialized universes, but that loses a lot of convenience.
* `UpdatingSpaceRaytracer` is not depended on in `all-is-cubes` and
  could be moved, but it depends on internals of `SpaceRaytracer`.