Deps: Use workspace inheritance for more dependencies.
I continue to have slightly mixed feelings about this, which is why I
haven’t migrated everything but it's useful to be able to view and edit
versions of everything in one place. The ones I haven't moved are the
more specifically “this is only going to be used by this one package”
libraries, like how only `all-is-cubes-desktop` uses `ratatui`. The ones
I have moved are the ones that are plausibly broadly useful utilities,
even if they're only currently used in one place. I am probably not
making sense.