ui: Improve dialog backgrounds.
Primarily, they now actually meet their contents, by being on the front
faces of their blocks; previously there was 15/16ths of a block of gap.

I also redrew the theme image. The new design has much less empty space
between the border and the content, is more rectangular, and has a faint
checkerboard pattern in the background. This checkerboard is not
necessarily the direction I want the art style to go, but as a temporary
measure, it will help when we’re debugging UI layout and want to count
how many cubes a widget is occupying, and it also adds some distraction
from widgets’ blocks’ shadows against the background.

In the future, I hope the dialog background will be able to be not just
2D, and will incorporate more shades of color. (Perhaps we will have a
complementary theme color that isn’t also the logo color.)