Add a `Schedule` to `Modifier::Move`.
I’m not entirely sure this is the right plan — will anyone want an
intentionally slower animation? Perhaps for performance, or perhaps not,
but what this does do is give a reference for the modifier’s `velocity`.
That is, anything with a `Schedule` in it is obviously something that
is dependent on the universe’s `TickSchedule`, and there's a systematic
way to update/validate those (at least, if we do something like
extending `VisitHandles` to visit other things).