Switch `[lints]` table syntax to not use dotted keys.
Dependabot’s parser currently doesn't like this for some reason
(<https://github.com/dependabot/dependabot-core/issues/10453>),
and while that’s certainly a bug, I’d like to get dependency updates
working again, and this is in fact the second time I’ve found a bug
in tools’ handling of dotted keys (the first time was
<https://github.com/toml-rs/toml/issues/673>), so it is pragmatic to
not do the thing that often doesn't work. For now, at least.

(The reasons I prefer the dotted keys are because one doesn’t have to
scan upward to the most recent table header to see whether the lint is
`rust` or `clippy`, and it’s more like `#[warn()]` syntax.)