gpu: Replace use of wgpu `.global_id()` with explicitly added IDs.
The next release of wgpu will remove `.global_id()`, so we have to stop
using them. It will also add `Eq` implementations, which could perhaps
substitute in some cases, but we usually want to compare resources which
currently exist with the IDs of resources that may be obsolete, so using
`Eq` would mean unnecessarily retaining those obsolete resources, which
is worse than generating our own IDs. So, this commit adds a wrapper
type `Identified` which essentially recreates the wgpu ID system.