content: Break `demo_city()` into smaller functions.
rustc has some sub-optimal handling of `async` functions and blocks, so
separating out code and variables that don't need to interact with the
asyncness helps keep the compiled code size, and in-memory size of the
`Future`s, smaller.

It doesn't seem to improve compilation time, unfortunately, but that
might be an effect too small to measure.

In the debug profile, the size of the `all-is-cubes` binary is reduced
by 29 kB, and in the release profile, by 3 kB. This is much less than
a percent of the whole binary, but for a change to essentially one
function, that's quite significant. I think I'll be taking a look at
other large async functions and blocks.