listen: Allow `Filter` to have a stack buffer for output batches.
I haven’t run benchmarks to confirm that this is actually worth doing.
I believe that either it is, or the whole message-batching system needs
rethinking — this is just finishing the implementation of the
current strategy.

Something we could perhaps do instead of this is create a notion of
message delivery “transactions” within which individual messages are
delivered one at a time, allowing taking a lock for the duration.
This way, it wouldn’t be necessary to compile messages into a slice
to achieve efficient locking; the disadvantage is that it exposes
more locking.