Add special docs.rs feature to enable all-is-cubes/std.
Without this, `all_is_cubes_mesh::dynamic::MeshJobQueue` is documented
as `!Send` when it ought to be conditionally `Send`.

The crates that need this extra feature are crates that don't already
enable all-is-cubes/std through any feature at all. all-is-cubes-mesh
Is the only one that actually meets that condition, but
all-is-cubes-content only depends indirectly in a way that I want to
remove, so I added the feature there too.

Also added an empty metadata section to all-is-cubes-ui, to remind
myself to fill it out as needed in the future.