mesh: Add no_std support by making the `dynamic` module optional.
I wrote that I wasn't going to bother doing this until I found a use
case, but it turns out it's really simple if we cut at the right spot.
And it makes sense to make the `dynamic` part optional since some
applications really are one-shot and have no need for it.

Also made the dep of `all-is-cubes-desktop` on `all-is-cubes-mesh`
optional since it is only used directly via `all-is-cubes-gpu`.