content: Use `Primitive::Text` in `make_some_voxel_blocks()`.
This removes one of the direct uses of `embedded_graphics`.
See <https://github.com/kpreid/all-is-cubes/issues/528>.

A disadvantage is that now evaluating the block is more expensive. We
could fix that by adding a `BlockDef` wrapper, or by making a way to
turn a `Block` evaluation into the contents of a `Space` (which we'll
probably want someday as an editing tool anyway).