Rename some files from `foo.rs` → `foo/mod.rs`.
I used to believe that the `foo.rs` style is superior because it
minimizes the number of duplicate file names in the project. However,
it has the significant disadvantage of making it harder to navigate
(in an ordinary file browser, including e.g. GitHub) from a child module
to the parent, except for those which are both outline-structured and
don't sort folders separately, and as the project has grown, there are
inevitably duplicate file names anyway, because the name which makes the
most *local* sense is not necessarily globally unique.

Furthermore, VS Code has an option to change the display name of files
by matching their pathnames, which can be used to ensure a `mod.rs` is
never displayed in isolation, and that file name search for `foo` will
automatically turn up `foo/mod.rs`:

```json
  "workbench.editor.customLabels.patterns": {
    "**/mod.rs": "${dirname}/mod.rs",
  },
```

Therefore, the IDE experience is no longer notably improved by distinct
names, and so the experience of working with naive file browsing is more
comparatively important than it was.

In this commit, I have only moved files which were largely composed of
`mod` and `use` items. Modules which contains significant items of their
own have been left alone, except for `all_is_cubes::universe`, which has
had `WhenceUniverse` split out of it, because that seemed like a
particularly good idea. In the future, I might do more such splits so
that the project style is more `mod.rs`-heavy and less mixed.