math: Replace `Geometry` trait with `Wireframe` trait and inherent `translate()` methods.
In practice, there were no *generic* uses of `Geometry::translate()`,
and it was often left unimplemented in order to gain access to the
wireframe method.

I looked at making other improvements to `wireframe_points()` at the
same time, but it turned out that they were not clearly better. I
replaced the TODO about making it return an iterator with an explanation
of why that would be worse. I considered making it return pairs of
points to avoid the error of unpaired points, but that seemed like it
would mostly add complexity to callers and callees, without great value.