listen: Batch messages passing through a `Filter` if they are ZSTs.
Ideally, we would always batch messages, but doing that with an
appropriate size of stack-allocated buffer would naïvely require the
caller to specify a size, which is a change with a larger impact. In
the ZST case we can just buffer up to `usize::MAX` messages for free.

(The `const {}` block probably doesn't do anything, but I put it there
to make it blatantly obvious to readers that the branch is statically
determined.)