block: Prototype for `Composite` merging operations.
The added code is more like what we *should* be doing, but still wrong.
Therefore, it is mostly disabled.

The big problem: this breaks `move_inside_composite_destination()`,
because there are two potential semantics for modifiers acting on ops:

1. modifiers modify the ops to stack on their effects, and
2. modifiers are usually inert; the originally provided op applies to
   the whole block,

and Move kind of assumed 2 while our new Composite is (necessarily?)
doing 1 in order to incorporate the source block.