Remove incorrect (I think) mentions of luminous emittance.
These were introduced in commit ffb394798db7f6c31978194630fd0e5a7ab2c984.
Thinking about these quantities continues to confuse me, but I'm pretty
sure that we mean luminance and not luminous emittance, because luminous
emittance is the light emitted from a surface *integrated over all
viewing directions*, i.e. the light flux passing through that boundary,
whereas the quantity we are concerned with is the light exiting the
surface in a single direction. These quantities are proportional as long
as we don’t support emission which varies with viewing angle (that is,
we are assuming every emitting material is a Lambertian emitter), but I
don’t want to introduce that coefficient of proportionality if I don’t
have to.