math: Use `PositiveSign` instead of `NotNan` in `Rgb` and `Rgba`.
It is now impossible for colors to contain negative components,
which will reduce the number of edge cases color processing code must
consider, and remove the the hazard of distinct negative zero
(<https://github.com/kpreid/all-is-cubes/issues/537>).

(There are some uses for negative-valued RGB components to represent
colors that are outside of the gamut of the chosen color space, but
I don’t know enough about that extension to know how to handle it
correctly.)

We’re probably also going to want a type restricted from 0 to 1, for
strict alpha and reflectance calculations, but `PositiveSign` is a
straightforward improvement of the overall situation.