Use `rstest` parameterized tests in some cases.
The advantage of this type of test (macro-generated test functions)
over the previous style is that each one can succeed or fail
independently. `all_is_cubes::util::MultiFailure` handles multiple
failures but doesn’t list them as separate tests. I’m not sure if
this is a good idea overall, though; it might be worse for build time
since there are more functions to compile, or it might be better than
`MultiFailure` since there is no closure.