render: `StandardCameras::update()` reports whether anything changed.
Arguably this should be signaled via a `Notifier` instead, but if we do
that, it would make the most sense to make `StandardCameras` itself a
sort of listenable thing that returns a struct of current state, instead
of its current combination of being both the data and the data source.
For now, this is what will be useful.