physics: Add `occupying` box to `Body` (not really used yet).
We currently define body collision in terms of `self.collision_box`
translated by `self.position`. The problem with this definition is that
rounding error means that it is difficult for collision detection to
achieve the specific absolute positions of the box's faces that
definitely don’t collide with any box, because after picking a value
for `self.position`, the resulting translated collision box then has
rounding error. `nudge_on_ray()` tries to fix this during collision
resolution, but doesn’t succeed sometimes.

I think we can solve this problem by defining a collision box in
absolute coordinates, which can be computed sometimes-exactly to
ensure they land on the correct side of surfaces collided with. As a
brute-force strategy, we can always slightly squish the box to
compensate for rounding errors that lead to intersection. Then, by
*also* keeping the original relative-coordinate collision box, we can
restore the intended shape when no collision occurs.

This approach will also help with “crushing” situations, where a body
stops having enough space for its intended box due to changes in the
`Space`, or when a body’s collision box is changed while it is already
resting against a surface, without teleporting the body to an arbitrary
location that might break intended game mechanics (as `push_out()`
currently does).

This commit does not implement any of the logic for these operations,
only add the `occupying` field and keep it up to date as a first step.