test-renderers: Delete `rend3` dependency and stub out `gltf-render`.
`rend3` is not being updated and is inconveniencing our dependency
updates, and the glTF tests never really worked right at all.
So, for now, I am making them entirely non-functional and dropping
the `rend3` dependency.

Arguably I should be deleting the tests entirely, but I hope to have a
chance of getting them back to sort-of-working soonish, just not quite
as quickly as I want the old deps gone.