test-renderers/gltf: Replace old rend3 code with somewhat newer Bevy code.
This gets to the point of rendering some scenes decently.
However, we don’t want to actually compile it yet, because it uses
Bevy 0.16 and that would duplicate `bevy_ecs` and transitive
dependencies. It is still on Bevy 0.16 because upgrading made it stop
working; troubleshooting that will require more work.

I’m merging this code now so that this code (which is an improvement on
its predecessor) can enter permanent version control, even though it
isn’t yet a functioning feature.

Besides the old dependency, there are several reasons this doesn’t
pass tests even if it were to be compiled:

* No textures (since our glTF exporter doesn’t have them).
* No smart light interpolation.
* No emissive materials.
* No info text.
* No non-uniform sky.
* No transparent volume materials.