gpu: Extract parts of `get_texels_from_gpu()`.
I am expecting this to be useful in future render testing work where
I need to do these operations with a device that isn't `Arc` wrapped in
the same way (specifically, replacing `rend3` with `bevy` in the render
tests of glTF output, though I haven't evaluated how good of an idea
that is regarding build costs). That may not succeed, but in any case,
this change is some to some extent a factoring into less-coupled pieces,
so it seems reasonable to do regardless.