listen: Split `all_is_cubes::listen` into a library, `nosy`.
This commit contains only the minimal migration changes; in particular,
`all_is_cubes::listen` still exists, and contains aliases for the old
item names rather than replacing them everywhere.

Rationale:

* Reduces the build time of `all-is-cubes` (both that crate in
  particular, and the whole project) by moving code into dependencies.
* Makes the functionality more available to other projects (my own and
  others) without taking a dependency on `all-is-cubes`.

A benefit of the migration is that the replacement of `ListenableSource`
with `DynSource` allows for sources that are neither `ListenableCell`
(now `Cell`) nor constants. We do not take advantage of this yet.