listen: Split `Listen::listen()` into provided and required methods.
This allows `dyn Listen` to exist and simplifies the code required in a
basic boilerplate `impl Listen`. The main motivation for it is that
I think there might be a better version of `ListenableSource` to be had
which is based on an `Arc<dyn SomethingThatHasListenAsASupertrait>` and
thus doesn't require the data source to be exactly a `ListenableCell`.