listen: Remove implicit `Arc` from `ListenableCell`.
This means we're not allocating `Arc<bool>`s or `Arc<Handle>`s.
It does add a fair bit of clutter of allocating `Arc`s, but it also
removes the subtly different `snapshot()` method and dereferencing of
`Copy` values. In the future we can also think about when the value in
the cell should itself be an `Arc`-like implicitly.

The real motivation, though, is that I want to split the listening logic
into a separate library, and it would be especially undesirable for that
library’s version of `ListenableSource` to make allocating an `Arc`
mandatory for all data communicated that way.