Light: Compile rays into a tree of shared prefixes.
Instead of separately traversing the path of each ray, we build a tree
structure that lets us avoid duplicating work in shared prefixes.

The goal of this work is to be able to afford more rays, but for now I
am keeping the same ray pattern (though there are some numerical
differences in the results); benchmarks show a 20-30% decrease in time.

Note that I have removed the build-time precalculation. This is because
the new chart is bigger (since it stores more information per step) and
will be multiple megabytes if made denser, and I don’t want that to be
contributing to the total binary size. I tried reducing the size by
deduplicating octants, but that turned out to be very hard to implement
correctly (I failed). All of this can be revisited later; the
`[FlatNode]` representation is friendly to being stored, provided that
we swap out `NonZero` and `FaceMap` for `bytemuck::Pod` versions.

(The original motivation for precalculation — which was added in commit
7b48879e4a008999553f8cb2b847e1017c4f14ce — wasn’t for faster startup,
but in order to remove lazy initialization, which I believed at the time
was necessary for `no_std` compatibility.)