gpu: Remove `Arc`s no longer necessary.
`wgpu` types now implement `Clone` to the same effect.

The one special case is that the polling system I previously had relied
on having a `Weak<wgpu::Device>`. I replaced it with a new design where
an explicit guard object is required to keep polling going. As a minor
theoretical benefit, it now stops polling in cases where the `Device`
still exists and has activity but explicit polling is not wanted.
That never happens in current practice, though.