block: Use a macro to derive things from `BlockAttributes`'s definition.
This reduces the amount of boilerplate code that has to be written for
each new attribute.

Note that not everything has been converted to the macro; in particular,
applying the macro to `rotate()` and serialization would want entire new
helper traits, which I decided not to do in this commit.