mesh: Compute mesh vertex positions during analysis.
This is not currently used, but I believe it will be necessary for
eventually fixing T-junctions, and it should currently compile to
nothing (when visualization is not enabled).

We will probably need to further refine this to include the directions
of edges, so that 2D polygon connectivity can be obtained; presuming
that the solution to T-junctions looks like using a conventional
polygon triangulation algorithm.