gpu: Increase `recursion_limit`.
This previously triggered with the update to wgpu 24 (81d0ffadc29dde4b78288b9a164d76a50af80609)
but it had only fired in `all-is-cubes-desktop` and not `all-is-cubes-gpu`.
I don’t know why it just became a problem now, but presumably it was
marginal, so increasing the limit in -gpu to match -desktop is reasonable.