mesh: Early exit if the block bounds are empty.
No test because I don’t know of a way to get block evaluation to actually
generate this case, and it only showed up via `EvaluatedBlock::arbitrary()`
in fuzzing. But, even if it shouldn’t happen, I still want to handle it,
on the general robustness principle of making functions total instead of
partial.