gpu: Simplify render passes.
`SpaceRenderer` creating its own render pass is both more complex than
necessary and less efficient. By creating the render passes in
`EverythingRenderer::draw_frame_linear()` instead, we can use the same
render pass for world and lines, remove a lot of parameters used to
configure the two calls to `SpaceRenderer::draw()`, and remove the
`output_needs_clearing` tracking.

This is not a significant performance improvement, as far as I know,
since it only reduces the number of passes from (4 + bloom stages)
to (3 + bloom stages), but if we ever have reason to render multiple
spaces as part of the world (e.g. independent islands in the same
coordinate system, or portals), it may make more of a difference.
For now, I’m just doing it for simplicity.