desktop: Use instanced blocks in glTF recording.
Note that this is not very efficient because it creates a glTF node per
instance. This is in fact all we can do with the base glTF specification;
instancing using data in buffers requires an extension
(<https://github.com/KhronosGroup/glTF/blob/main/extensions/2.0/Vendor/EXT_mesh_gpu_instancing/README.md>)
and I currently do not intend to implement that extension until I have
an actual use cases for the glTF export where that extension is
definitely going to be supported by the recipient.