gpu: Move `write_camera_only()` call from `update()` to `draw()`.
This makes the CPU raytracing mode more parallel with non-raytracing,
because in the non-raytracing mode `SpaceRenderer::draw()`, not
`update()`, is responsible for the buffer. Our general principle is
that `update()` copies the state of the world, but `draw()` can pick
whatever camera state it wants to draw, except for the view distance
limit applying to the camera position from the update.

Also update some comments to be more accurate and condense some code.