gpu: Use `StagingBelt` for camera and instance buffer writes.
This is to minimize per-frame allocations and copies.

Previously I was thinking that this wasn’t feasible because the belt
was already in the “finished” state, but in fact it’s perfectly fine
to do multiple `write()` and `finish()` as long as the encoders are
submitted before `recall()`. We just have to use an extra command
encoder to let the belt commands be mixed with the render pass, but
that's not an extra *submission*, so it should be cheap.