ui: Move `InventoryWatcher` up so it can be shared.
This will be useful when we have an inventory-viewing page, because they
will be able to share the same `InventoryWatcher`.

Eventually, this should belong to some kind of "reactive" system for
scheduling the UI update work and be able to be automatically
created as needed, instead of being a special field and being separately
updated by every widget instance.