ui: Generalize `Toolbar` widget to work on a range of slots.
This is a first step towards having a proper inventory page.
It also implements the TODO about clearing nonexistent positions.

Development note: This logic deserves a unit test (particularly for the
cases where it avoids panicking on out-of-range slots), but I had some
trouble writing the test, and will revisit that.