ui: Redefine `HudInputs` and `Vui` to use `UiTargets` as much as possible.
The goal of this change is to minimize the amount of additional plumbing
that is required to add a new data source to the UI as a whole (all the
way from `Session` to individual pages).

Arguably, `HudInputs` should go away and be replaced with narrower
parameter passing, but for now I just want the code to be easier to edit
when creating new UI features. Is this a good idea? I don’t know, but I
don’t think the old version was particularly *better*.