ui: Use images to define button shapes.
Benefits:

* Removes a usage of `embedded_graphics` (which I would like to remove
  entirely; see <https://github.com/kpreid/all-is-cubes/issues/528>).

* Reduces the binary size by 3 kB — compressed PNGs are smaller than
  drawing algorithms!

* When I decide to tweak the button shapes or make new ones (e.g.
  segmented buttons) they can be just drawn instead of coded.

One thing I’m not entirely confident in here is (implicitly) using a
full nine-patch image instead of stretching the individual button
shapes — there's a lot of redundant pixels in the images. But, it
would be more code to add that stretching, so let’s stick with this
until we have a reason to change it.