raytracer: Add 1 ray bounce (with multiple samples).
This is controlled by the new variant `LightingOption::Bounce`.

Just 1 bounce makes a massive improvement in lighting fidelity, because
our existing per-cube light suffices to figure out what the incoming
light should be quite well, but it works less well for points within the
viewed block.

Of course, it’s still much slower and has noise. In order to use this more
seriously, we would want to add accumulation across frames or use it only
for non-realtime renders.