vui: Add "draw" function to widgets.
This new trait method is currently called but never overridden
to do anything. Eventually, all widgets will be expected to be able to
redraw themselves on command, so that we can do more dynamic UIs that
hide or rearrange widgets rather than having a fixed layout for an
entire `Space` of widgets.