dungeon: Shrink rooms to 7×7.
Right now, I find it tedious to walk the dungeon without flying, which
would be an obstacle to testing once I start making chasms be actual
game obstacles (i.e. stop supplying the jetpack as a starting item).
This should help reduce that tedium, and make the rooms less notably
empty, though I do want to add more greebles and features like
challenges to overcome within each room. (The *original* plan for the
dungeon was that it would form a tutorial for All is Cubes controls
and demonstration of what kind of game mechanics it can support, not
just be a maze with items.)