render: Make `HeadlessRenderer::update()` not async.
Neither implementation takes advantage of the async signature,
nor do any of the callers, and I don’t think any will. In particular,
it’s not very useful for the caller to be stuck with the constraint
“the universe is still potentially being read until this future
completes” — most main loops aren’t themselves async in their
individual steps. If we want renderers to do asynchronous update work,
more useful approaches would be background jobs or a third method to be
called between `update()` and `draw()`.

More concretely, I am experimenting with stricter rules about
`Universe` access, and I want to be able to declare that the access
is always done within a single synchronous step.