Add `ReadTicket`, a parameter to almost everything that reads from a `Universe`.
This parameter serves two roles: it documents the dependencies which
could create a run-time borrow conflict if not handled appropriately,
and if in the future we decide to switch to a different data storage
implementation where a `Handle` alone is not sufficient to obtain access
to data (such as an ECS where `Handle`s become entity IDs; see
<https://github.com/kpreid/all-is-cubes/issues/620>), it will carry the
actual borrow or other temporary access.

Places where further work is probably necessary are marked by use of
`ReadTicket::new()`, or `// TODO(read_ticket)`.

The largest, most commonly used, such unhandled case is `Space::set()`
and friends. It would be highly annoying to have to pass a `ReadTicket`
for every operation that mutates a `Space`. However, such mutations are
primarily worldgen (and tests) and I think that it may be feasible to
approach them using some kind of batch access API: make a struct that
combines a `&mut Space` and a `ReadTicket`, then provide get/set
operations on that. Such a struct could also optimize writes by
offering a way to pre-register blocks and write by index — and remove
fallibility of individual operations, because unlike adding a `Block`
to a `Space`, changing the index in a particular cube cannot fail except
if the index is outright invalid. (But all that is just one speculative
idea for how to handle things.)