space: Make it possible to execute `SpaceTransaction` on `MutationCtx`.
I am planning to make `MutationCtx` (possibly under another name)
public and use it broadly for worldgen. This will ensure the necessary
`ReadTicket` is present, and allow notification batching, and maybe
other uses like setting `BlockIndex` and skipping redundant `Block`
comparisons. But in order to do that, we need to be able to execute
a `SpaceTransaction` since many worldgen functions have been written
to use transactions.