space: Add `Builder::read_ticket()`.
This handles providing a `ReadTicket` when using `filled_with()`, and
will be useful in further changes.

The default `ReadTicket` used is now `ReadTicket::stub()`, so callers
have to supply the ticket if using blocks that refer to a universe.
(Or more precisely, if the ticket is bad, the block evaluations will
be wrong until the next step. This is bad, but existing behavior, which
we should probably revisit later.)