raytracer: Change `Accumulate::add()` to take a struct instead of separate parameters.
Note that this has the cost that it is no longer possible for arbitrary
code to call `Accumulate::add()` freely. I don’t see a way to do that
while preserving the designed extensibility, though, other than adding
constructors to `Hit` (which might be good once we have more of the
space of possibilities).