gpu: Pass depth from raytracer to GPU depth buffer.
This allows the rasterized lines to be occluded by the raytraced scene.
I plan to also use this depth information for reprojection to produce a
more stable raytraced image.

Note that the depth values are not correct for UI, because UI has a
different camera and coordinate system. In order to fix that, we will
need to either pass the UI and world as separate textures instead of
combining them into one (higher data rate), or have a flag per pixel
identifying which source the pixel is from (doesn't trace areas of
the world hidden by the UI in the *current* camera position).