gpu: raytrace_to_texture: Add implementation of reprojection.
Upside: Smooth camera movement.
Downside: lots of distracting black gaps.

This very much needs further work, so it is disabled and not actually
used. (We really need a proper configuration system to let the user set
all these little extra cases.)

Note that the conventional real-time raytracing reprojection strategy
is backwards rather than forwards: rather than drawing points based on
the previous frame as we are doing here, the fragment shader would
project the *current* frame's pixel position back to its position in the
previous frame. The reason we are not doing that is because we are not
(yet) trying to accumulate samples over multiple historical frames, but
to extrapolate future frames we haven’t traced yet at all, so we have
no fresh depth value with which to reproject backwards.