universe: Change tests that use unreferenced anonymous members to use named members.
These usages would break when GC actually follows its rules and collects
anonymous members not referred to by named members.

Eventually we will probably want a way to make a “strong reference”
handle that works even when the handle is not used within the `Universe`,
but that is not yet definitely needed. For now, I'm doing this for the
sake of the refactor to ECS-style storage, where it makes a lot more
sense to do real GC rather than looking at `Handle` reference counts.