universe: Add `StrongHandle`.
This currently does nothing, but eventually it will ensure that
`Universe` garbage collection (which currently is fake, in that it
actually just checks reference counts) does not collect universe members
that are in use externally. Some of those uses may eventually become
ordinary handles inside of the `Universe`, but this will, at a minimum,
help with the migration.

The specific places `StrongHandle` is now used are:
* `Session`'s current character
* `StandardCameras`'s current character
* `PageInst`'s `Space`