universe: Store handles’ names atomically.
I am hoping that further changes along these lines may allow replacing
the state lock entirely, and perhaps even making handles always `Sync`.

It also has the nice effect that `Handle::name()` doesn't need to take
the state lock (which has been a source of deadlocks when I call it from
something that holds the state lock already). In principle, it could be
changed to return `&Name` instead of cloning `Name`, but I’ll hold off
for now as that's a wider impact.