port: Make `export_to_path()`’s asyncness more useful.
Instead of the future borrowing the `ReadTicket`, all reads are
performed immediately. The async block contains only as much of the
work as does not require the ticket.

Further improvement could be made by returning two futures for the two
parts, allowing the caller to multitask even the reading part if
desired, but that will require even more refactoring (each format’s
exporter will need to stuff its data into a channel between the two
tasks).