universe: Remove ability to create pending handles from `Arbitrary`.
Callers wishing to use `impl Arbitrary for Handle` must now use the new
struct `ArbitraryWithUniverse`, which sets up a thread-local context in
the same way as is used for deserialization. This is awkward, but the
`arbitrary` feature is mostly for our own fuzz testing, so I’m not too
worried.

This change also revealed that `Character::spawn_default()` has an
`unwrap()` that it really shouldn’t. That should be fixed separately.