gpu: raytrace_to_texture: Avoid `PixelPicker when not needed.
We do this by introducing `UpdateStrategy`, an enum containing
either a `PixelPicker` or a simple counter.

Also don't collect an unnecessary `Vec<Point>` in single-threaded
execution. This means that there are four cases instead of two, but
I’ve arranged to keep the code in each case as small as possible.

These changes improve performance when it is most needed.