gpu: raytrace_to_texture: Add filling in the gaps between reprojected points.
This is a very sloppy and unstable technique but does a pretty good job
as a *temporary* fill into areas that otherwise would have nothing at
all and thus would previously be black. It could use further work:

* Don’t mix UI pixels into the world (this would need either separate
  rendering for each or identification to entirely reject the UI data).
* Compare depths to avoid background data blending into the foreground.

These additional features are the reason why I’ve modified `mip_ping`
to support multiple textures, but I haven’t made use of that yet.

This commit has no effect if the `UpdateStrategy` is not set to use
reprojection.