desktop: Add “reload” command.
When clicked (from the pause menu) it re-runs the `UniverseSource`
load or generation so that you get a fresh copy like relaunching
the game.

The intended use case of this command is to allow “hot”-reloading
universe templates. Currently, it just reruns the same code so it only
resets the world state unless the current source is a file that was
modified.

Future work:
* Actually run fresh template code, by spawning a `cargo run` that
  writes a save file that we read (only in development builds, of
  course).
* Preserve character’s position.
* Automatically unpause.
* Move this UI component into the `ui` crate so that it is available to
  the web build too.