Lint: Fix future `clippy::manual_is_multiple_of` lint.
I'm not entirely comfortable with using `is_multiple_of(2)` to express
"do X if even and Y if odd", because it makes more asymmetry, but I
suppose `i % 2 == 0` isn't especially clear either. The ideal case would
be being able to write something like

    match value % 2 {
        0 => ...
        1 => ...
    }

but that doesn’t count as an exhaustive match.