mesh: Extract new function `IndexVec::extend_with_offset()`.
This eliminates duplicated code in `write_block_mesh_to_space_mesh`.
Despite the `inline(always)` (to preserve previous performance), this
also produces a reduction in binary size of 8 kB (on aarch64).

As noted in the comments, it is also possible that we could make
`extend_with_offset` smarter where `impl Extend for IndexVec` can’t be,
but that’s a complex change which I am putting off for now.